Chatbot/requirements.txt
fastapi
uvicorn[standard]
requests
geocoder
python-dotenv
pymongo
pandas
numpy
scikit-learn
spacy
openai
google-generativeai
yake
spacy
